Handle large country ids

+++ src/vic3_world/characters/vic3_country_character_map_importer.cpp
@@ -17,7 +17,7 @@ std::map<int, std::vector<int>> vic3::ImportCountryCharacterMap(std::istream& in
    commonItems::parser map_parser;
    map_parser.registerRegex(commonItems::integerRegex,
        [&country_character_map, &characters](const std::string& number_string, std::istream& input_stream) {
-          const int country_id = std::stoi(number_string);
+          const int country_id = static_cast<int>(std::stoul(number_string));
           country_character_map.emplace(country_id, commonItems::getInts(input_stream));
           characters += static_cast<int>(country_character_map.at(country_id).size());
        });
